pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observatorio_geosalud"
version = "0.1.0"
description = "Análisis descriptivo espacial de dengue — Valle del Cauca, Colombia"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Germán Avila", email = "germanavila09@gmail.com"}
]
keywords = [
    "dengue", "epidemiology", "geospatial", "gis",
    "machine-learning", "time-series", "forecasting"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
]
requires-python = ">=3.10"
dependencies = [
    "pandas>=2.0",
    "geopandas>=1.1.3",
    "numpy>=1.26",
    "matplotlib>=3.10.8",
    "seaborn>=0.13.2",
    "SQLAlchemy>=2.0.49",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "neuralprophet>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=3.0",
    "black>=22.0",
    "flake8>=4.0",
    "mypy>=0.950",
    "jupyter>=1.0",
    "jupyterlab>=3.0",
]

[project.urls]
Homepage = "https://github.com/germanavila09/Observatorio-GeoSalud"
Documentation = "https://github.com/germanavila09/Observatorio-GeoSalud"
Repository = "https://github.com/germanavila09/Observatorio-GeoSalud.git"
Issues = "https://github.com/germanavila09/Observatorio-GeoSalud/issues"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ['py310', 'py311', 'py312']
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
  | outputs
  | lightning_logs
)/
'''

[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
addopts = "-v --strict-markers"
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow tests",
]

[tool.coverage.run]
source = ["src"]
omit = [
    "*/tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
